Use scrollbar API in text windows.
Former-commit-id: 167abab232cdb9f9a42098c69f2b805d155154ad

+++ client/src/include/widget.h
@@ -58,7 +58,7 @@ typedef struct widgetdata
 	int show;
 
 	/** Widget must be redrawn. */
-	int redraw;
+	uint8 redraw;
 
 	/** Should this widget be the only one of its type? */
 	int unique;
